[GR-34785] Order implicit casts from most frequent to least frequent

+++ src/main/java/org/truffleruby/language/NoImplicitCastsToLong.java
@@ -16,13 +16,17 @@ import com.oracle.truffle.api.dsl.TypeSystem;
 @TypeSystem
 public abstract class NoImplicitCastsToLong {
 
+    // Ordered from most frequent to least frequent for interpreter performance
+
+    // For handling interop primitives
+
     @ImplicitCast
-    public static int promoteToInt(byte value) {
+    public static int promoteToInt(short value) {
         return value;
     }
 
     @ImplicitCast
-    public static int promoteToInt(short value) {
+    public static int promoteToInt(byte value) {
         return value;
     }
